+++ src/apprt/gtk/App.zig
@@ -967,7 +967,7 @@ fn showDesktopNotification(
     defer notification.unref();
     notification.setBody(n.body);
 
-    const icon = gio.ThemedIcon.new("com.mitchellh.ghostty");
+    const icon = gio.ThemedIcon.new("com.ghostkellz.ghostshell");
     defer icon.unref();
     notification.setIcon(icon.as(gio.Icon));